+++ src/main/java/com/mmz/specs/service/DetailListService.java
@@ -27,14 +27,14 @@ public interface DetailListService {
 
     public void setDetailListDao(DetailListDao detailListDao);
 
-    public int addDetailList(DetailListEntity detailListEntity);
+    public long addDetailList(DetailListEntity detailListEntity);
 
     public void updateDetailList(DetailListEntity detailListEntity);
 
-    public void removeDetailList(int id);
+    public void removeDetailList(long id);
 
 
-    public DetailListEntity getDetailListById(int id);
+    public DetailListEntity getDetailListById(long id);
 
 
     public List<DetailListEntity> getDetailListByParent(DetailEntity detailEntity);
